+++ src/service/links/queries.ts
@@ -4,10 +4,11 @@ import LinkService from '@/service/links/linkService';
 //create unique key
 const queryKeys = {
   //get method
-  link: ({ linkId }: { linkId: number }) => {
-    return ['link', linkId] as const;
+  link: ({ accessToken, linkId }: { accessToken: string; linkId: number }) => {
+    return ['link', accessToken, linkId] as const;
   },
-  reminder: ['reminder'] as const,
+  reminder: ({ accessToken }: { accessToken: string }) =>
+    ['reminder', accessToken] as const,
 
   //post method
   newLink: ['newLink'] as const,
@@ -34,24 +35,26 @@ const errorHandler = (error: any) => {
 
 const queryOptions = {
   //get method
-  link: ({ linkId }: { linkId: number }) => ({
-    queryKey: queryKeys.link({ linkId }),
-    queryFn: () => LinkService.getLink({ linkId }),
+  link: ({ accessToken, linkId }: { accessToken: string; linkId: number }) => ({
+    queryKey: queryKeys.link({ accessToken, linkId }),
+    queryFn: () => LinkService.getLink({ accessToken, linkId }),
     onError: errorHandler,
   }),
-  linkReminder: () => ({
-    queryKey: queryKeys.reminder,
-    queryFn: () => LinkService.getLinkReminder(),
+  linkReminder: ({ accessToken }: { accessToken: string }) => ({
+    queryKey: queryKeys.reminder({ accessToken }),
+    queryFn: () => LinkService.getLinkReminder({ accessToken }),
     onError: errorHandler,
   }),
   //post method
   newLink: ({
+    accessToken,
     autoComplete,
     title,
     summary,
     url,
     categoryId,
   }: {
+    accessToken: string;
     autoComplete: boolean;
     title: string;
     summary: string;
@@ -60,16 +63,25 @@ const queryOptions = {
   }) => ({
     queryKey: queryKeys.newLink,
     queryFn: () =>
-      LinkService.postLink({ autoComplete, title, summary, url, categoryId }),
+      LinkService.postLink({
+        accessToken,
+        autoComplete,
+        title,
+        summary,
+        url,
+        categoryId,
+      }),
     onError: errorHandler,
   }),
   updateLink: ({
+    accessToken,
     title,
     summary,
     url,
     categoryId,
     linkId,
   }: {
+    accessToken: string;
     title: string;
     summary: string;
     url: string;
@@ -78,18 +90,37 @@ const queryOptions = {
   }) => ({
     queryKey: queryKeys.updateLink({ linkId }),
     queryFn: () =>
-      LinkService.updateLink({ title, summary, url, categoryId, linkId }),
+      LinkService.updateLink({
+        accessToken,
+        title,
+        summary,
+        url,
+        categoryId,
+        linkId,
+      }),
     onError: errorHandler,
   }),
-  updateStatus: ({ linkId }: { linkId: number }) => ({
+  updateStatus: ({
+    accessToken,
+    linkId,
+  }: {
+    accessToken: string;
+    linkId: number;
+  }) => ({
     queryKey: queryKeys.updateStatus({ linkId }),
-    queryFn: () => LinkService.updateLinkStatus({ linkId }),
+    queryFn: () => LinkService.updateLinkStatus({ accessToken, linkId }),
     onError: errorHandler,
   }),
   //delete method
-  deleteLink: ({ linkId }: { linkId: number }) => ({
+  deleteLink: ({
+    accessToken,
+    linkId,
+  }: {
+    accessToken: string;
+    linkId: number;
+  }) => ({
     queryKey: queryKeys.deleteLink({ linkId }),
-    queryFn: () => LinkService.deleteLink({ linkId }),
+    queryFn: () => LinkService.deleteLink({ accessToken, linkId }),
     onError: errorHandler,
   }),
 };
